apis/yamlgen/requirements.txt
PyYAML>=6.0.1
pathlib>=1.0.1 